requirements.txt
aiogram>=3.20.0,<4.0.0
aiosqlite>=0.20.0,<1.0.0
openpyxl>=3.1.5,<4.0.0
python-dotenv>=1.0.1,<2.0.0
tzdata>=2025.2
